tests/requirements.txt
pytest>=7.4.0
hypothesis>=6.82.0
pyyaml>=6.0
